add new tool nlplug-findfs
nlplug-findfs is a tool that will help us to find a given filesystem
using coldplug. The only reason that initramfs exist in first place is
to set up a rootfs and switch to it. The nlplug-findfs will handle
coldplug events til a given filesystem/device is found and then mount
it.