nlplug-findfs: avoid lstat syscall wen scanning /sys if possible
There are ~10k entries in /sys so we want be fast. Not all filesystems
supports dirent filed d_type, (iso9660 for example) so we need use lstat
to find directories. However, we know that sysfs supports it, so we can
avoid 10k lstat syscalls. This might be noticiable difference on rpi.