nlplug-findfs: make sure we handle all events in queue
we want handle all the events in queue, even if we find the devices we
are looking for. This is so we load kernel module for framebuffer even
if the apkvol and bootrepo are found first.

This means that the only valid exit condition is a poll timeout and that
we can set the timeout to 0 once we found all the things we are looking
for.