mkinitfs: add support for resume
This patch tries to add resume from hibernation (aka suspend to disk)
support to mkinitfs. The swap partition can be provided via the resume
option.

I know this feature is mostly needed on destops/notebooks, therefore I'm
not sure if there is any interest for Alpine Linux. The default Alpine
Linux
kernel has no hibernation support compiled in, so you need to build
a custom kernel in order to test this.

These links helped me on the topic:
* https://wiki.gentoo.org/wiki/Custom_Initramfs/Hibernation
* https://git.archlinux.org/mkinitcpio.git/tree/hooks/resume

As this might be a dangerous feature I would like to get your opinions.