mkinitfs: add support for resume
This patch tries to add resume from hibernation (aka suspend to disk)
support to mkinitfs. The swap partition can be provided via the resume
option.

The default Alpine Linux kernel has no hibernation support compiled in,
so a custom kernel is needed in order to use this.

Some links on the topic:
* https://wiki.gentoo.org/wiki/Custom_Initramfs/Hibernation
* https://git.archlinux.org/mkinitcpio.git/tree/hooks/resume