initramfs-init: add support for s390x installer
On s390x, the boot media consist of a kernel, an initramfs, and a parm
file. We don't necessarily need a modloop image and initial packages
but fetch from remote repo instead. This behavior is the same on other
distros.

The SSH server is required to allow root access with password during
installation. This config will be removed after running setup-disk.
User sets password in the parm file along with network interface
types, IP addresses, DNS server, disks.