Add archfile var and use proper getty for ppc64le
On ppc64le when net botting Alpine, we will not get a login prompt from
Petitboot. This patch will set TTY as /dev/console and the TERMTYPE as
vt100 while using a safer BAUD value in order work on varying ppc64le
hardware. Additionally I create a archfile variable which is also used
to silently fix the apk arch in case the apkovl does not match.