initramfs-init: Always configure network when ip= is given
Previously, only alpine_repo= and apkovl= triggered the configuration
of the network. Not using them caused the ip= parameter to be ignored.

The invocations of configure_ip for said two options were not removed
because network configuration may still be triggered by the options
named above. The function itself is safe to be invoked multiple times.