Reuse kernel-side configuration for console= devices
Instead of parsing the console= options in the initramfs, we ask the
kernel of its interpretation of the console= parameters. The kernel
does the console setup as part of its early startup, including the
configuration of the baud rate, control bits and flow control.

The options and format of the console= parameter are documented here:
https://www.kernel.org/doc/html/v4.15/admin-guide/serial-console.html

By keeping the settings from the kernel, we avoid baud rate switching
between printk and getty output on edge cases.

This adds support for additional tty types, like hvc, while removing
string parsing code from the initramfs.