nlplug-findfs: detect zfs pool
if search device is prefixed with ZFS= then we search for a label with
the zpool name in the zfs path. For example, if search device is
"ZFS=tank/alpine/root" then we search for device that is type
"zfs_member" and label "tank".

This makes it work better with grub which creates a boot cmdline with
ZFS=