add feature rpirtc
Add new feature rpi rtc allowing a hw rtc to be used.

The init script in Alpine Linux since v3.9 looks for /dev/rtc, if
not found it will switch to swclock.

To make this check work on a Rasberry PI with a mounted rtc and
the following in usercfg.txt
 dtoverlay=i2c-rtc,ds3231
we must have rtc drivers available already initramfs.