init: create cryptsetup locking directory
The LUKS2 cryptsetup format uses a locking mechanism, see the "LUKS2
header locking" section in cryptsetup(8) for further information on this
mechanism. The mechanism requires a locking directory, if it is not
present it will be created automatically. Unfortunately, cryptsetup
emits a warning when creating it.

To silence this warning message make init create the directory.