mkinitfs: add optional support for pigz
pigz(1) is a parallel gzip implementation. As such, it significantly
faster than gzip(1). This patch uses pigz(1) instead of gzip(1),
if it is installed, otherwise it just falls back to gzip(1). This is
similar to the current pigz support in abuild.

Without this patch:

	$ time mkinitfs
	==> initramfs: creating /boot/initramfs-lts
	real	0m 38.19s
	user	0m 35.47s
	sys	0m 0.49s

With this patch applied and pigz installed:

	==> initramfs: creating /boot/initramfs-lts
	real	0m 13.24s
	user	0m 38.30s
	sys	0m 0.56s

This change was discussed on IRC with _ikke_ and mps.