initramfs-init: fix importing the root pool
We accidentally used $_root_vol instead of $_root_pol for force importing.
Also import (but without -f) if we don't force import as this is required
on some setups. Add -d /dev as this is required for zpool to find the
pool if you have defined some subvolume as partition (e.g. for VMs)