Add zstd compression support
Now that 5.10 is out and supports zstd compression for the kernel and
initramfs, add support for zstd to mkinitfs. The benefits have been
discussed at length on lkml[1][2].

In my personal experience, the compression (mkinitfs) seems to be
slower, but decompression (booting) seems to be faster. I think most
people boot more than they mkinitfs.

The default is still gzip, so no change should be noticed unless
someone goes looking for trouble.

To test this functionality, just add the following to mkinitfs.conf:

```text
initfscomp=zstd
```

[1] https://lkml.org/lkml/2020/7/23/1010
[2] https://lkml.org/lkml/2020/7/1/722