mkinitfs.in: ensure when run in chroot that correct kernel version is used
modinfo is called without the "-k version" argument and so when
running mkinitfs inside a chroot where the version of the kernel
package being installed differs from the version of the host
machine's kernel then modinfo is called (to check firmware
requirements) using the wrong kernel version.

Add the "-k" argument to ensure the correct kernel version is used.

NOTE: Busybox's modinfo does not support the "-k" option. Currently
Alpine is using modinfo from the kmod package, however obviously in
the future Busybox's version is to be used this this will break. It
appears the only solution in that scenario would be to add "-k"
support to Busybox's version.