init: support for booting from an aoe device
- aoe: causes the aoe module to be loaded. if passed a comma separated
list of targets, it will attempt to explicitly discover these targets.
aoe autodiscovery is automatic on linux, so passing a list of targets
should not generally be needed.

- aoe_iflist: takes a comma separated list of ethernet interfaces to
attach to the aoe subsystem. this also ensures that those links are set
to `up'. you effectively must pass this to use aoe because it is useless
without being bound to at least one interface.

it is up to the user to add the `network' and `aoe' features to
mkinitfs.conf, and the necessary modules for the nic to the bootloader
modules= configuration.

Signed-off-by: grobe0ba <grobe0ba@tcp80.org>