mkinitfs: Support compressed firmware
When executing modinfo, the provided firmware names do not consider
whether the on-disk files are compressed or not. As such, we need
to manually check for this in order to support compressed firmware.

Tested by running $ sudo ../mkinitfs -o test-compressed.gz, then
unzipping the result and checking the file list using $ cpio -t.
The resulting file list did indeed contain the compressed firmware.

As an extra precaution I also did the same steps with uncompressed
firmware, and that code path still works as well.