init: fix apkovl on another device/fs
Partially fixes #74.

Omitting /dev/ prefix is removed due to a lack of flexibility. As
my intention is to use virtiofs whose `device' parameter is not an
actual path, support for /dev/mapper/<name> mentioned in the issue
can also be implemented more easily.

Added a basic test of this feature.